mcp_server/requirements.txt
# MCP server (HTTP/SSE transport via FastMCP)
mcp>=1.2
uvicorn>=0.30
starlette>=0.37
httpx>=0.27
python-dotenv>=1.0
